+++ src/main/java/insane96mcp/iguanatweaksreborn/module/experience/feature/OtherExperience.java
@@ -6,14 +6,24 @@ import insane96mcp.insanelib.base.Label;
 import insane96mcp.insanelib.base.Module;
 import insane96mcp.insanelib.base.config.Config;
 import insane96mcp.insanelib.base.config.LoadFeature;
+import net.minecraft.ChatFormatting;
+import net.minecraft.nbt.CompoundTag;
+import net.minecraft.network.chat.Component;
 import net.minecraft.server.level.ServerLevel;
 import net.minecraft.world.entity.ExperienceOrb;
 import net.minecraft.world.entity.projectile.ThrownExperienceBottle;
 import net.minecraft.world.item.ItemStack;
+import net.minecraft.world.item.Items;
+import net.minecraft.world.item.enchantment.Enchantment;
 import net.minecraft.world.item.enchantment.EnchantmentHelper;
 import net.minecraft.world.item.enchantment.Enchantments;
-import net.minecraftforge.event.entity.player.AnvilRepairEvent;
+import net.minecraftforge.api.distmarker.Dist;
+import net.minecraftforge.api.distmarker.OnlyIn;
+import net.minecraftforge.event.AnvilUpdateEvent;
+import net.minecraftforge.event.entity.player.ItemTooltipEvent;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
+
+import java.util.Map;
 
 @Label(name = "Other Experience", description = "Change other experience sources")
 @LoadFeature(module = Modules.Ids.EXPERIENCE)
@@ -28,11 +38,14 @@ public class OtherExperience extends Feature {
 	@Label(name = "Remove rename cost", description = "Removes cost of renaming items in Anvil")
 	public static Boolean freeRenaming = true;
 	@Config
-	@Label(name = "Unmending", description = "Replaces the default Mending enchantment. Mending sets the repair cost of an item to 'Unmending Cap' and will stop it from increasing. No longer repairs items with xp.")
+	@Label(name = "Unmending", description = """
+			Makes mending reset the repair cost of an item to 0 when applied to it. No longer repairs items with XP.
+			If an item has already mending, the enchantment will be removed and repair cost reset.
+			Applying mending still requires the base repair cost of the item (you can't add Mending if the operation is 'Too Expensive'""")
 	public static Boolean unmending = true;
-	@Config(min = 1)
-	@Label(name = "Unmending Cap", description = "Set the cap repair cost set by Unmending")
-	public static Integer unmendingCap = 25;
+	@Config
+	@Label(name = "Repaired Tooltip", description = "If true (and Unmending is enabled), items will have a 'Item has been repaired' tooltip.")
+	public static Boolean repairedTooltip = false;
 
 	public OtherExperience(Module module, boolean enabledByDefault, boolean canBeDisabled) {
 		super(module, enabledByDefault, canBeDisabled);
@@ -53,14 +66,65 @@ public class OtherExperience extends Feature {
 	}
 
 	@SubscribeEvent
-	public void onAnvilUse(AnvilRepairEvent event) {
+	public void onAnvilUse(AnvilUpdateEvent event) {
 		if (!this.isEnabled()
 				|| !unmending)
 			return;
-		ItemStack output = event.getOutput();
-		if (EnchantmentHelper.getItemEnchantmentLevel(Enchantments.MENDING, output) > 0 && output.getBaseRepairCost() > 15) {
-			output.setRepairCost(unmendingCap);
+
+		ItemStack left = event.getLeft();
+		ItemStack right = event.getRight();
+		ItemStack out = event.getOutput();
+
+		if(out.isEmpty() && (left.isEmpty() || right.isEmpty()))
+			return;
+
+		boolean isMended = false;
+
+		Map<Enchantment, Integer> enchLeft = EnchantmentHelper.getEnchantments(left);
+		Map<Enchantment, Integer> enchRight = EnchantmentHelper.getEnchantments(right);
+
+		if(enchLeft.containsKey(Enchantments.MENDING) || enchRight.containsKey(Enchantments.MENDING)) {
+			if(left.getItem() == right.getItem())
+				isMended = true;
+
+			if(right.getItem() == Items.ENCHANTED_BOOK)
+				isMended = true;
 		}
+
+		if(isMended) {
+			if(out.isEmpty())
+				out = left.copy();
+
+			if(!out.hasTag())
+				out.setTag(new CompoundTag());
+
+			Map<Enchantment, Integer> enchOutput = EnchantmentHelper.getEnchantments(out);
+			enchOutput.putAll(enchRight);
+			enchOutput.remove(Enchantments.MENDING);
+
+			EnchantmentHelper.setEnchantments(enchOutput, out);
+
+			out.setRepairCost(0);
+			if(out.isDamageableItem())
+				out.setDamageValue(0);
+
+			event.setOutput(out);
+			if(event.getCost() == 0)
+				event.setCost(1);
+		}
+	}
+
+	@SubscribeEvent
+	@OnlyIn(Dist.CLIENT)
+	public void onTooltip(ItemTooltipEvent event) {
+		if (!this.isEnabled()
+				|| !unmending
+				|| !repairedTooltip)
+			return;
+
+		int repairCost = event.getItemStack().getBaseRepairCost();
+		if(repairCost > 0)
+			event.getToolTip().add(Component.translatable("iguanatweaksreborn.item_repaired").withStyle(ChatFormatting.YELLOW));
 	}
 
 	public static boolean isUnmendingEnabled() {