+++ changelog.md
@@ -3,6 +3,7 @@
 ## Upcoming
 * Requires InsaneLib 1.7.1
 * Added back Client module
+* Unmending now resets the repair cost of an item instead of capping the repair cost.
 
 ## Alpha 2.13.0
 * Updated to 1.19.1+